Cleanup with the ring code
Versions are now int64s, essentially just time.Now.UnixNano().

Capacities are now int32s to allow sums of them to fit in int64s.

Renamed some Impl to not have that suffix if not really needed.

Renamed TierIDs to TierSeparations for a bit more clarity.

Shortened a couple internal method names for easier reading.

Dropped the whole interfacing of RingBuilder. The Ring is the part
that's most useful as an interface.