Pretty good amount of work...
Major CLI work. Should meet requirements for now; more niceties later.

Tier string values are now managed/persisted/loaded.

I keep flip-flopping on this, but Ring and Node are now interfaces
again. This is protect Nodes from having their internals changed which
would have affected how a Builder builds rings. It also gives a bit
easier testing for Node/Ring mockups. There's a separate BuilderNode
this time, for changing node attributes for the Builder to use.

Changed adding a node to a builder to require all the details rather
than a node instance (since the builder is the only one who creates the
instances now).

Added ability to change replica count (not in CLI yet).

Getting a ring from a builder now returns the ring with no local node
binding. Just seems easier, for testing certainly, and it's easy enough
to SetLocalNode on the ring afterwards if desired.

Unexported a lot of items that really didn't need to be.