Stop connection loop...
Stop connection loop if the underlying msgChan is no longer the same or
is no longer set. This was specifically occurring when the loop could
not connect to an addr and was in a retry loop and then a ring change
came in removing that node; the retry loop would just keep going.